+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - When using http endpoint to access the service, you can now pass workspace specific token to the http header `Authorization` to access the service. (Previously, all the services are assumed to be accessed from the same service provider workspace)
  - Breaking Change: Remove `info`, `warning`, `error`, `critical`, `debug` from the `hypha` module, use `log` or `log_event` instead.
  - Support basic observability for the workspace, including workspace status, event bus and websocket connection status.
+ - Support download statistics for the artifacts in the artifact manager.
 
 ### 0.20.37
  - Add s3-proxy to allow accessing s3 presigned url in case the s3 server is not directly accessible. Use `--enable-s3-proxy` to enable the s3 proxy when starting Hypha.